fix: unify hooks in one
