+++ packages/react-components/react-motion-preview/src/hooks/index.ts
@@ -1,4 +1,2 @@
 export * from './useMotion';
-export { getDefaultMotionState } from './useMotionPresence';
-export type { MotionOptions, MotionState, MotionType } from './useMotionPresence';
 export * from './useMotionClassNames';
