--- packages/react-components/react-motion-preview/src/hooks/useMotionDuration.test.ts
@@ -1,162 +1,0 @@
-import { act, renderHook } from '@testing-library/react-hooks';
-
-import { useMotionDuration, MotionShorthand, UseMotionDurationOptions } from './useMotionDuration';
-import { getDefaultMotionState } from './useMotionPresence';
-
-const cssDuration = 100;
-const renderHookWithRef = (
-  initialMotion: MotionShorthand,
-  initialOptions?: UseMotionDurationOptions,
-  style: Record<string, string | undefined> = { 'transition-duration': `${cssDuration}ms` },
-) => {
-  const refEl = document.createElement('div');
-  const hook = renderHook(({ motion, options }) => useMotionDuration(motion, options), {
-    initialProps: {
-      motion: initialMotion,
-      options: initialOptions,
-    } as { motion: MotionShorthand; options?: UseMotionDurationOptions },
-  });
-
-  Object.entries(style).forEach(([key, value]) => value && refEl.style.setProperty(key, value));
-
-  function renderRef() {
-    act(() => {
-      if (!hook.result.current.canRender) {
-        return;
-      }
-
-      if (typeof hook.result.current.ref === 'function') {
-        hook.result.current.ref(refEl);
-      }
-    });
-  }
-
-  act(() => renderRef());
-
-  function rerender(motion: MotionShorthand, options?: UseMotionDurationOptions) {
-    hook.rerender({ motion, options });
-    act(() => renderRef());
-  }
-
-  return { ...hook, rerender };
-};
-
-const jumpAnimationFrame = () => {
-  // requestAnimationFrame
-  act(() => jest.advanceTimersToNextTimer());
-};
-
-const jumpAnimationTimeout = (timeout: number = 300) => {
-  // timeout + requestAnimationFrame
-  act(() => {
-    jest.advanceTimersByTime(timeout);
-    jest.advanceTimersToNextTimer();
-  });
-};
-
-describe('useMotion', () => {
-  let computedStyleMock: jest.SpyInstance;
-
-  beforeEach(() => {
-    jest.useFakeTimers();
-    computedStyleMock = jest.spyOn(window, 'getComputedStyle').mockImplementation(() => {
-      return {
-        getPropertyValue: () => `${cssDuration}ms`,
-      } as unknown as CSSStyleDeclaration;
-    });
-  });
-
-  afterEach(() => {
-    jest.useRealTimers();
-    jest.resetAllMocks();
-  });
-
-  describe('when presence is boolean', () => {
-    it('should sync presence value with canRender', () => {
-      const { result, rerender } = renderHookWithRef(false);
-
-      expect(result.current.canRender).toBe(false);
-      rerender(true);
-
-      expect(result.current.canRender).toBe(true);
-    });
-
-    it('should return default values when presence is false', () => {
-      const { result } = renderHookWithRef(false);
-
-      expect(typeof result.current.ref).toBe('function');
-      expect(result.current.type).toBe('unmounted');
-      expect(result.current.active).toBe(false);
-      expect(result.current.canRender).toBe(false);
-    });
-
-    it('should return default values when presence is true', () => {
-      const { result } = renderHookWithRef(true);
-
-      expect(typeof result.current.ref).toBe('function');
-      expect(result.current.active).toBe(true);
-      expect(result.current.type).not.toBe('unmounted');
-      expect(result.current.canRender).toBe(true);
-    });
-
-    it('should change visible to true when animateOnFirstMount is true', () => {
-      const { result } = renderHookWithRef(true, { animateOnFirstMount: true });
-
-      expect(typeof result.current.ref).toBe('function');
-      expect(result.current.active).toBe(false);
-      expect(result.current.canRender).toBe(true);
-
-      act(() => jest.advanceTimersToNextTimer());
-
-      expect(result.current.active).toBe(true);
-    });
-  });
-
-  describe('when duration is provided', () => {
-    it('should only call getComputedStyle when duration is not provided', () => {
-      const { result, rerender } = renderHookWithRef(false, { duration: 300 });
-
-      expect(typeof result.current.ref).toBe('function');
-      expect(result.current.type).toBe('unmounted');
-      expect(result.current.active).toBe(false);
-      expect(result.current.canRender).toBe(false);
-
-      rerender(true, { duration: 300 });
-
-      expect(result.current.canRender).toBe(true);
-      expect(result.current.type).toBe('entering');
-      expect(result.current.active).toBe(false);
-
-      jumpAnimationFrame();
-      jumpAnimationTimeout();
-
-      expect(computedStyleMock).not.toHaveBeenCalled();
-
-      rerender(false);
-
-      jumpAnimationFrame();
-      jumpAnimationTimeout();
-
-      expect(computedStyleMock).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  describe('when presence is a MotionShorthand', () => {
-    it('should return default values when presence is false', () => {
-      const defaultState = getDefaultMotionState();
-      const { result } = renderHookWithRef(defaultState);
-
-      expect(result.current.type).toStrictEqual('unmounted');
-      expect(result.current.ref).toStrictEqual(defaultState.ref);
-      expect(result.current.active).toStrictEqual(false);
-    });
-
-    it('should return default values when presence is true', () => {
-      const defaultState = getDefaultMotionState();
-      const { result } = renderHookWithRef({ ...defaultState, active: true });
-
-      expect(result.current.ref).toStrictEqual(defaultState.ref);
-      expect(result.current.active).toStrictEqual(true);
-    });
-  });
-});
